+++ src/game/Arenas.h
@@ -59,8 +59,8 @@ public:
 
 	int32 GetFreeTeam()
 	{
-		uint32 c0 = m_players[0].size() + m_pendPlayers[0].size();
-		uint32 c1 = m_players[1].size() + m_pendPlayers[1].size();
+		size_t c0 = m_players[0].size() + m_pendPlayers[0].size();
+		size_t c1 = m_players[1].size() + m_pendPlayers[1].size();
 		if(m_started) return -1;
 
 		// Check if there is free room, if yes, return team with less members
